+++ pypsa/linopf.py
@@ -1540,7 +1540,7 @@ def assign_solution(
 def network_lopf(
     n,
     snapshots=None,
-    solver_name="cbc",
+    solver_name="glpk",
     solver_logfile=None,
     extra_functionality=None,
     multi_investment_periods=False,
@@ -1567,7 +1567,7 @@ def network_lopf(
         network.snapshots, defaults to network.snapshots
     solver_name : string
         Must be a solver name that pyomo recognises and that is
-        installed, e.g. "glpk", "gurobi"
+        installed, e.g. "highs", "glpk", "gurobi"
     solver_logfile : None|string
         If not None, sets the logfile option of the solver.
     solver_options : dictionary