+++ pypsa/opf.py
@@ -2281,7 +2281,7 @@ def network_lopf_prepare_solver(network, solver_name="glpk", solver_io=None):
     ----------
     solver_name : string
         Must be a solver name that pyomo recognises and that is
-        installed, e.g. "glpk", "gurobi"
+        installed, e.g. "highs", "glpk", "gurobi"
     solver_io : string, default None
         Solver Input-Output option, e.g. "python" to use "gurobipy" for
         solver_name="gurobi"
@@ -2443,7 +2443,7 @@ def network_lopf(
         network.snapshots, defaults to network.snapshots
     solver_name : string
         Must be a solver name that pyomo recognises and that is
-        installed, e.g. "glpk", "gurobi"
+        installed, e.g. "highs", "glpk", "gurobi"
     solver_io : string, default None
         Solver Input-Output option, e.g. "python" to use "gurobipy" for
         solver_name="gurobi"